Coupled user with a profile, implemented change-password api and added security on todos api

+++ src/main/java/com/todo/rest/webservices/restfulwebservices/user/UserDAO.java
@@ -3,8 +3,10 @@ package com.todo.rest.webservices.restfulwebservices.user;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface UserDAO extends CrudRepository<DAOUser, Integer> {
+public interface UserDAO extends CrudRepository<DAOUser, Long> {
     DAOUser findByUsername(String username);
 
 }
